Remove pending protocol schemas
