+++ packages/protocol/types.ts
@@ -12,6 +12,7 @@ import type {
   ActResultSchema,
   AnthropicModelIdSchema,
   AnthropicModelNameSchema,
+  ApiKeyAuthSchema,
   AzureEntraIdAuthSchema,
   AzureModelProviderOptionsSchema,
   AzureProviderOptionsSchema,
@@ -29,6 +30,8 @@ import type {
   CachingSchema,
   CerebrasModelIdSchema,
   CerebrasModelNameSchema,
+  ClientOptionsBaseSchema,
+  ClientOptionsSchema,
   ClientModelReferenceSchema,
   ClearCookieOptionsSchema,
   ContextActivePageResultSchema,
@@ -110,6 +113,7 @@ import type {
   LLMStructuredGenerateResultSchema,
   LLMTextContentSchema,
   LLMTextResponseFormatSchema,
+  LLMToolSchema,
   LLMToolAnnotationsSchema,
   LLMToolChoiceSchema,
   LLMToolExecutionSchema,
@@ -117,9 +121,12 @@ import type {
   LLMToolResultContentSchema,
   LLMToolUseContentSchema,
   LLMUsageSchema,
+  LocalBrowserLaunchOptionsSchema,
   MouseButtonSchema,
+  ModelAuthSchema,
   ModelConfigSchema,
   ModelNameSchema,
+  ModelProviderOptionsSchema,
   ModelProviderSchema,
   GroqModelIdSchema,
   GroqModelNameSchema,
@@ -180,6 +187,7 @@ import type {
   StagehandPingResultSchema,
   SnapshotResultSchema,
   TelemetryConfigSchema,
+  ThinkingEffortSchema,
   OpenAIModelIdSchema,
   OpenAIModelNameSchema,
   VariablePrimitiveSchema,
@@ -188,64 +196,6 @@ import type {
   VertexModelProviderOptionsSchema,
   VertexProviderOptionsSchema,
 } from "./schemas.js";
-import type {
-  ActRequestSchema,
-  ActResponseSchema,
-  AISDKApiKeyProviderSchema,
-  AnthropicClientOptionsSchema,
-  ApiKeyAuthSchema,
-  ApiKeyClientOptionsSchema,
-  ApiKeyResolvedProviderClientOptionsSchema,
-  AzureApiKeyClientOptionsSchema,
-  AzureEntraClientOptionsSchema,
-  AzureResolvedProviderClientOptionsSchema,
-  BrowserConfigSchema,
-  ClientOptionsBaseSchema,
-  ClientOptionsSchema,
-  ClipboardOptionsSchema,
-  ClipboardPasteOptionsSchema,
-  ErrorResponseSchema,
-  ExtractRequestSchema,
-  ExtractResponseSchema,
-  HistoryEntrySchema,
-  LLMToolSchema,
-  LocalBrowserLaunchOptionsSchema,
-  ModelAuthSchema,
-  ModelProviderOptionsSchema,
-  NavigateOptionsSchema,
-  NavigateRequestSchema,
-  NavigateResponseSchema,
-  NavigateResultSchema,
-  ObserveRequestSchema,
-  ObserveResponseSchema,
-  OllamaResolvedProviderClientOptionsSchema,
-  OpenAIClientOptionsSchema,
-  ReplayActionSchema,
-  ReplayPageSchema,
-  ReplayResponseSchema,
-  ReplayResultSchema,
-  ResolvedProviderClientOptionsSchema,
-  SessionEndResponseSchema,
-  SessionEndRequestSchema,
-  SessionEndResultSchema,
-  SessionHeadersSchema,
-  SessionIdParamsSchema,
-  SessionStartRequestSchema,
-  SessionStartResponseSchema,
-  SessionStartResultSchema,
-  StreamEventLogDataSchema,
-  StreamEventSchema,
-  StreamEventStatusSchema,
-  StreamEventSystemDataSchema,
-  StreamEventTypeSchema,
-  ThinkingEffortSchema,
-  TokenUsageSchema,
-  V3FunctionNameSchema,
-  VertexClientOptionsSchema,
-  VertexResolvedProviderClientOptionsSchema,
-  defaultExtractSchema,
-  pageTextSchema,
-} from "./pending-schemas.js";
 
 export type VariablePrimitive = z.infer<typeof VariablePrimitiveSchema>;
 export type VariableValue = z.infer<typeof VariableValueSchema>;
@@ -415,18 +365,7 @@ export type StagehandRpcNotification = z.infer<typeof StagehandRpcNotificationSc
 export type StagehandMethod = z.infer<typeof StagehandMethodSchema>;
 export type StagehandSendToHostBinding = z.infer<typeof StagehandSendToHostBindingSchema>;
 
-export type AnthropicClientOptions = z.infer<typeof AnthropicClientOptionsSchema>;
-export type AISDKApiKeyProvider = z.infer<typeof AISDKApiKeyProviderSchema>;
 export type ApiKeyAuth = z.infer<typeof ApiKeyAuthSchema>;
-export type ApiKeyClientOptions = z.infer<typeof ApiKeyClientOptionsSchema>;
-export type ApiKeyResolvedProviderClientOptions = z.infer<
-  typeof ApiKeyResolvedProviderClientOptionsSchema
->;
-export type AzureApiKeyClientOptions = z.infer<typeof AzureApiKeyClientOptionsSchema>;
-export type AzureEntraClientOptions = z.infer<typeof AzureEntraClientOptionsSchema>;
-export type AzureResolvedProviderClientOptions = z.infer<
-  typeof AzureResolvedProviderClientOptionsSchema
->;
 export type BrowserbaseBrowserSource = z.infer<typeof BrowserbaseBrowserSourceSchema>;
 export type BrowserbaseRegion = z.infer<typeof BrowserbaseRegionSchema>;
 export type BrowserbaseSessionCreateParams = z.infer<typeof BrowserbaseSessionCreateParamsSchema>;
@@ -434,35 +373,18 @@ export type Caching = z.infer<typeof CachingSchema>;
 export type ClearCookieOptions = z.infer<typeof ClearCookieOptionsSchema>;
 export type ClientOptions = z.infer<typeof ClientOptionsSchema>;
 export type ClientOptionsBase = z.infer<typeof ClientOptionsBaseSchema>;
-export type ClipboardOptions = z.infer<typeof ClipboardOptionsSchema>;
-export type ClipboardPasteOptions = z.infer<typeof ClipboardPasteOptionsSchema>;
 export type Cookie = z.infer<typeof CookieSchema>;
 export type CookieFilter = z.infer<typeof CookieFilterSchema>;
 export type CookieParam = z.infer<typeof CookieParamSchema>;
 export type CookieRegex = z.infer<typeof CookieRegexSchema>;
 export type DomainPolicy = z.infer<typeof DomainPolicySchema>;
-export type ErrorResponse = z.infer<typeof ErrorResponseSchema>;
-export type HistoryEntry = z.infer<typeof HistoryEntrySchema>;
 export type LLMTool = z.infer<typeof LLMToolSchema>;
 export type LoadState = z.infer<typeof LoadStateSchema>;
 export type LocalBrowserLaunchOptions = z.infer<typeof LocalBrowserLaunchOptionsSchema>;
 export type ModelAuth = z.infer<typeof ModelAuthSchema>;
 export type ModelProviderOptions = z.infer<typeof ModelProviderOptionsSchema>;
-export type OllamaResolvedProviderClientOptions = z.infer<
-  typeof OllamaResolvedProviderClientOptionsSchema
->;
-export type OpenAIClientOptions = z.infer<typeof OpenAIClientOptionsSchema>;
-export type ResolvedProviderClientOptions = z.infer<typeof ResolvedProviderClientOptionsSchema>;
 export type ThinkingEffort = z.infer<typeof ThinkingEffortSchema>;
-export type V3FunctionName = z.infer<typeof V3FunctionNameSchema>;
-export type VertexClientOptions = z.infer<typeof VertexClientOptionsSchema>;
-export type VertexResolvedProviderClientOptions = z.infer<
-  typeof VertexResolvedProviderClientOptionsSchema
->;
 
-export type ActRequest = z.infer<typeof ActRequestSchema>;
-export type ActResponse = z.infer<typeof ActResponseSchema>;
-export type BrowserConfig = z.infer<typeof BrowserConfigSchema>;
 export type BrowserbaseBrowserSettings = z.infer<typeof BrowserbaseBrowserSettingsSchema>;
 export type BrowserbaseContext = z.infer<typeof BrowserbaseContextSchema>;
 export type BrowserbaseFingerprint = z.infer<typeof BrowserbaseFingerprintSchema>;
@@ -471,32 +393,4 @@ export type BrowserbaseProxyConfig = z.infer<typeof BrowserbaseProxyConfigSchema
 export type BrowserbaseProxyGeolocation = z.infer<typeof BrowserbaseProxyGeolocationSchema>;
 export type BrowserbaseViewport = z.infer<typeof BrowserbaseViewportSchema>;
 export type ExternalProxyConfig = z.infer<typeof ExternalProxyConfigSchema>;
-export type ExtractRequest = z.infer<typeof ExtractRequestSchema>;
-export type ExtractResponse = z.infer<typeof ExtractResponseSchema>;
-export type NavigateRequest = z.infer<typeof NavigateRequestSchema>;
-export type NavigateOptions = z.infer<typeof NavigateOptionsSchema>;
-export type NavigateResponse = z.infer<typeof NavigateResponseSchema>;
-export type NavigateResult = z.infer<typeof NavigateResultSchema>;
-export type ObserveRequest = z.infer<typeof ObserveRequestSchema>;
-export type ObserveResponse = z.infer<typeof ObserveResponseSchema>;
 export type ProxyConfig = z.infer<typeof ProxyConfigSchema>;
-export type ReplayAction = z.infer<typeof ReplayActionSchema>;
-export type ReplayPage = z.infer<typeof ReplayPageSchema>;
-export type ReplayResponse = z.infer<typeof ReplayResponseSchema>;
-export type ReplayResult = z.infer<typeof ReplayResultSchema>;
-export type SessionEndResponse = z.infer<typeof SessionEndResponseSchema>;
-export type SessionEndRequest = z.infer<typeof SessionEndRequestSchema>;
-export type SessionEndResult = z.infer<typeof SessionEndResultSchema>;
-export type SessionHeaders = z.infer<typeof SessionHeadersSchema>;
-export type SessionIdParams = z.infer<typeof SessionIdParamsSchema>;
-export type SessionStartRequest = z.infer<typeof SessionStartRequestSchema>;
-export type SessionStartResponse = z.infer<typeof SessionStartResponseSchema>;
-export type SessionStartResult = z.infer<typeof SessionStartResultSchema>;
-export type StreamEvent = z.infer<typeof StreamEventSchema>;
-export type StreamEventLogData = z.infer<typeof StreamEventLogDataSchema>;
-export type StreamEventStatus = z.infer<typeof StreamEventStatusSchema>;
-export type StreamEventSystemData = z.infer<typeof StreamEventSystemDataSchema>;
-export type StreamEventType = z.infer<typeof StreamEventTypeSchema>;
-export type TokenUsage = z.infer<typeof TokenUsageSchema>;
-export type DefaultExtract = z.infer<typeof defaultExtractSchema>;
-export type PageText = z.infer<typeof pageTextSchema>;
