+++ packages/server/llm/LLMProvider.ts
@@ -1,6 +1,6 @@
 import type { LanguageModel, LanguageModelMiddleware } from "ai";
 import type { ClientOptions, ModelName, ModelProvider } from "../../protocol/types.js";
-import { ClientOptionsSchema, ModelNameSchema } from "../../protocol/pending-schemas.js";
+import { ClientOptionsSchema, ModelNameSchema } from "../../protocol/schemas.js";
 import { AISdkClient } from "./aisdk.js";
 import { LLMClient } from "./LLMClient.js";
 import { createOpenAI } from "@ai-sdk/openai";
